[GSoC 2026] Refactor: split chatbot tool serializers/tests into per-tool packages (refs #3732)
