+++ api_app/chatbot_manager/serializers/__init__.py
@@ -1,0 +1,2 @@
+# This file is a part of IntelOwl https://github.com/intelowlproject/IntelOwl
+# See the file 'LICENSE' for copying permission.
